undo count result types switch from long to int - that should be another PR

+++ crates/mongodb-agent-common/src/query/response.rs
@@ -222,10 +222,10 @@ fn type_for_aggregates(query_aggregates: &IndexMap<ndc_models::FieldName, Aggreg
         .map(|(field_name, aggregate)| {
             let result_type = match aggregate {
                 Aggregate::ColumnCount { .. } => {
-                    Type::Scalar(MongoScalarType::Bson(mongodb_support::BsonScalarType::Long))
+                    Type::Scalar(MongoScalarType::Bson(mongodb_support::BsonScalarType::Int))
                 }
                 Aggregate::StarCount => {
-                    Type::Scalar(MongoScalarType::Bson(mongodb_support::BsonScalarType::Long))
+                    Type::Scalar(MongoScalarType::Bson(mongodb_support::BsonScalarType::Int))
                 }
                 Aggregate::SingleColumn { result_type, .. } => result_type.clone(),
             };
